+++ src/ext/java/org/opentripplanner/ext/vectortiles/PropertyMapper.java
@@ -5,7 +5,6 @@ import com.wdtinc.mapbox_vector_tile.adapt.jts.IUserDataConverter;
 import com.wdtinc.mapbox_vector_tile.adapt.jts.UserDataKeyValueMapConverter;
 import com.wdtinc.mapbox_vector_tile.build.MvtLayerProps;
 import java.util.Collection;
-import org.opentripplanner.common.model.T2;
 
 /**
  * This class is used for adding data for each object in the vector layer from the userData in the
@@ -22,15 +21,12 @@ public abstract class PropertyMapper<T> implements IUserDataConverter {
     VectorTile.Tile.Feature.Builder featureBuilder
   ) {
     if (userData != null) {
-      for (T2<String, Object> e : map((T) userData)) {
-        final String key = e.first;
-        final Object value = e.second;
-
-        if (key != null && value != null) {
-          final int valueIndex = layerProps.addValue(value);
+      for (var e : map((T) userData)) {
+        if (e.key() != null && e.value() != null) {
+          final int valueIndex = layerProps.addValue(e.value());
 
           if (valueIndex >= 0) {
-            featureBuilder.addTags(layerProps.addKey(key));
+            featureBuilder.addTags(layerProps.addKey(e.key()));
             featureBuilder.addTags(valueIndex);
           }
         }
@@ -41,5 +37,5 @@ public abstract class PropertyMapper<T> implements IUserDataConverter {
   /**
    * The return type is to allow null values.
    */
-  protected abstract Collection<T2<String, Object>> map(T input);
+  protected abstract Collection<KeyValue> map(T input);
 }