+++ src/ext/java/org/opentripplanner/ext/vectortiles/layers/vehicleparkings/DigitransitVehicleParkingPropertyMapper.java
@@ -5,8 +5,8 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Locale;
 import javax.annotation.Nonnull;
-import org.opentripplanner.common.model.T2;
 import org.opentripplanner.ext.vectortiles.I18NStringMapper;
+import org.opentripplanner.ext.vectortiles.KeyValue;
 import org.opentripplanner.ext.vectortiles.PropertyMapper;
 import org.opentripplanner.routing.vehicle_parking.VehicleParking;
 
@@ -23,23 +23,23 @@ public class DigitransitVehicleParkingPropertyMapper extends PropertyMapper<Vehi
   }
 
   @Override
-  protected Collection<T2<String, Object>> map(VehicleParking vehicleParking) {
+  protected Collection<KeyValue> map(VehicleParking vehicleParking) {
     return basicMapping(vehicleParking);
   }
 
   @Nonnull
-  protected ArrayList<T2<String, Object>> basicMapping(VehicleParking vehicleParking) {
+  protected ArrayList<KeyValue> basicMapping(VehicleParking vehicleParking) {
     return new ArrayList<>(
       List.of(
-        new T2<>("id", vehicleParking.getId().toString()),
-        new T2<>("bicyclePlaces", vehicleParking.hasBicyclePlaces()),
-        new T2<>("anyCarPlaces", vehicleParking.hasAnyCarPlaces()),
-        new T2<>("carPlaces", vehicleParking.hasCarPlaces()),
-        new T2<>(
+        new KeyValue("id", vehicleParking.getId().toString()),
+        new KeyValue("bicyclePlaces", vehicleParking.hasBicyclePlaces()),
+        new KeyValue("anyCarPlaces", vehicleParking.hasAnyCarPlaces()),
+        new KeyValue("carPlaces", vehicleParking.hasCarPlaces()),
+        new KeyValue(
           "wheelchairAccessibleCarPlaces",
           vehicleParking.hasWheelchairAccessibleCarPlaces()
         ),
-        new T2<>("name", i18NStringMapper.mapToApi(vehicleParking.getName()))
+        new KeyValue("name", i18NStringMapper.mapToApi(vehicleParking.getName()))
       )
     );
   }