+++ src/main/java/org/opentripplanner/api/model/ApiAgency.java
@@ -5,22 +5,13 @@ import java.util.Objects;
 
 public class ApiAgency implements Serializable {
 
-  private static final long serialVersionUID = 1L;
-
   public String id;
-
   public String name;
-
   public String url;
-
   public String timezone;
-
   public String lang;
-
   public String phone;
-
   public String fareUrl;
-
   public String brandingUrl;
 
   @Override