+++ src/main/java/org/opentripplanner/standalone/config/buildconfig/DemConfig.java
@@ -1,6 +1,7 @@
 package org.opentripplanner.standalone.config.buildconfig;
 
 import static org.opentripplanner.standalone.config.framework.json.OtpVersion.V2_2;
+import static org.opentripplanner.standalone.config.framework.json.OtpVersion.V2_3;
 
 import org.opentripplanner.graph_builder.module.ned.parameter.DemExtractParameters;
 import org.opentripplanner.graph_builder.module.ned.parameter.DemExtractParametersBuilder;
@@ -12,7 +13,21 @@ import org.opentripplanner.standalone.config.framework.json.NodeAdapter;
  */
 public class DemConfig {
 
-  public static DemExtractParametersList mapDemConfig(NodeAdapter root, String parameterName) {
+  public static DemExtractParameters mapDemDefaultsConfig(NodeAdapter root, String parameterName) {
+    var baseDefaults = DemExtractParameters.DEFAULT;
+    var demDefaults = root
+      .of(parameterName)
+      .since(V2_3)
+      .summary("Default properties for DEM extracts.")
+      .asObject();
+    return mapGenericDemParameters(demDefaults, baseDefaults, "").build();
+  }
+
+  public static DemExtractParametersList mapDemConfig(
+    NodeAdapter root,
+    String parameterName,
+    DemExtractParameters defaults
+  ) {
     return new DemExtractParametersList(
       root
         .of(parameterName)
@@ -29,31 +44,45 @@ public class DemConfig {
           the command line.
           """
         )
-        .asObjects(DemConfig::mapDemExtract)
+        .asObjects(nodeAdapter -> mapDemExtract(nodeAdapter, defaults))
     );
   }
 
-  private static DemExtractParameters mapDemExtract(NodeAdapter config) {
-    return new DemExtractParametersBuilder()
+  private static DemExtractParameters mapDemExtract(
+    NodeAdapter config,
+    DemExtractParameters defaults
+  ) {
+    String documentationAddition = " Overrides the value specified in `demDefaults`.";
+    return mapGenericDemParameters(config, defaults, documentationAddition)
       .withSource(
         config.of("source").since(V2_2).summary("The unique URI pointing to the data file.").asUri()
       )
+      .build();
+  }
+
+  private static DemExtractParametersBuilder mapGenericDemParameters(
+    NodeAdapter node,
+    DemExtractParameters defaults,
+    String documentationAddition
+  ) {
+    return defaults
+      .copyOf()
       .withElevationUnitMultiplier(
-        config
+        node
           .of("elevationUnitMultiplier")
-          .since(V2_2)
-          .summary("Specify a multiplier to convert elevation units from source to meters.")
+          .since(V2_3)
+          .summary(
+            "Specify a multiplier to convert elevation units from source to meters." +
+            documentationAddition
+          )
           .description(
             """
             Unit conversion multiplier for elevation values. No conversion needed if the elevation
             values are defined in meters in the source data. If, for example, decimetres are used
-            in the source data, this should be set to 0.1. This overrides the value specified in
-            [`elevationUnitMultiplier`](#elevationUnitMultiplier) in the build config at root level.
-          """
+            in the source data, this should be set to 0.1.
+            """
           )
-          .asDoubleOptional()
-          .orElse(null)
-      )
-      .build();
+          .asDouble(defaults.elevationUnitMultiplier())
+      );
   }
 }