+++ src/main/java/org/opentripplanner/transit/service/StopModelBuilder.java
@@ -1,6 +1,7 @@
 package org.opentripplanner.transit.service;
 
 import org.opentripplanner.framework.collection.CollectionsView;
+import org.opentripplanner.framework.statistics.MedianCalcForDoubles;
 import org.opentripplanner.transit.model.basic.WgsCoordinate;
 import org.opentripplanner.transit.model.framework.EntityById;
 import org.opentripplanner.transit.model.site.AreaStop;
@@ -9,7 +10,6 @@ import org.opentripplanner.transit.model.site.GroupStop;
 import org.opentripplanner.transit.model.site.MultiModalStation;
 import org.opentripplanner.transit.model.site.RegularStop;
 import org.opentripplanner.transit.model.site.Station;
-import org.opentripplanner.util.MedianCalcForDoubles;
 
 public class StopModelBuilder {